requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==2.0.12
Django==3.2
idna==3.4
newsapi-python==0.2.7
python-coinmarketcap==0.5
requests==2.26.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==1.26.16
attrs==23.1.0
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
isort==5.12.0
packaging==23.1
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2023.3
rest-framework-simplejwt==0.0.2
toml==0.10.2
six==1.16.0
sorl-thumbnail==12.7.0
Faker==12.0.1